pyproject.toml
[project]
name = "awesome-windows-applications"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bbcode>=1.1.0",
    "jinja2>=3.1.6",
    "minify-html>=0.18.1",
    "pyyaml>=6.0.3",
    "tzdata>=2025.3",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
